refactor(dishes): rename module
